♻️  F12: Passed code duplication in the context DRY
